ci: Refactor the TestPropertiesPlugin into generic PropertiesPlugin
The PropertiesPlugin is used to read the local.properties or any
other properties file and add them to the project extras or to
create new properties files.

Properties can be set either via the `local.properties` file or
the gradle cli using `-P...` arguments.

Currently it's used by the Health API Library to create the
`test.properties` file and by the Health SDK example app to create
the `client.properties` file.