ci: Store the version in each module's `gradle.properties` file
Using only tags to set the version is unfortunately not possible. The
injected version overrides all `project` dependencies' versions, too.

Say we have module B which has a project dependency on module A.
When we release module B with version 3.0.0, then in the POM the
dependency on module A will have version 3.0.0, too. Even though
module A has no version 3.0.0.

To fix this we need to update the version in the module's
`gradle.properties` and commit it to git before each release.