feat(core-api-library): Extract shared api lib code to the new Core API Library
The Core API Library is meant to be used as an internal dependency
for our API libraries only. It should not be used by our clients
directly. It contains public code, but clients may use it only through
one of our API libraries.

Its builder does not set a `GiniApiType` implicitly. One has to be set
when used by our API libraries.

The `GiniApiType.DEFAULT` has been deprecated in favor of industry
specific APIs. It is used in tests though.