ci: Merge dokka with used sibling modules
Dokka 1.5.30 added support for multi-module gradle projects. It can
merge the documentation for all child projects. It ignores dependencies
between child projects and does not merge the documentation of child
projects with the sibling projects they depend on.

Added a sibling collector task for generating html and javadoc
documentations. It merges the documentation only with sibling projects
that are used in the dependencies.