fix(health-api-library): Get page preview using Health API's `/pages` endpoint
The Health API has no direct endpoint to get a page preview image.

We need to get all the pages and then find the page image url
for the required resolution ourselves.

We have a maximum resolution of 4 MP and pick the largest or equal
available resolution. If all resolutions are greater then we pick
the smallest available one.