feat(capture-sdk): Allow injecting top and bottom navigation bar views
Injection of custom views happens through a view provider interface called `InjectedViewProvider`.
Implementations of this interface provide our SDK with the view instance.

The GiniCapture builder allows setting custom `InjectedViewProvider` implementation instances.

In this commit we allow setting implementations of the `NavigationBarTopProvider` and
`NavigationBarBottomProvider`. Both extend the `InjectedViewProvider`.

The view instance will be then shown in our layouts via a custom layout implementation called
`InjectedViewContainer`.

In this commit the OnboardingFragment has an `InjectedViewContainer` for the top and one for the
bottom navigation bar.

The bottom navigation bar can be enabled via `setBottomNavigationBarEnabled()` when building a
GiniCapture instance.

PIA-2290