feat(capture-sdk): Allow injecting custom views for the onboarding illustrations
The `OnboardingIconProvider` can be used to inject custom views for the onboarding illustrations.
The `onVisible()` and `onHidden()` methods are useful when using animations.

The `GiniCapture` builder allows setting custom implementations for each onboarding page.

Also changed the default pages according to the new UI design, but this is still work in progress.

PIA-2290