refactor(capture-sdk): Use the term `adapter` instead of `provider` for injected views
Provider was misleading because our interface is not only a factory, but it also needs to hold on to
the injected view to fulfill the interface contract requirements (e.g., to set a click listener or
a title).
The adapter term is used widely in Android (for RecyclerView, ListView, etc.) for view factories
with additional methods affecting the created views.

PIA-2290