refactor(capture-sdk): Rename `InjectedViewAdapter.getView()` to `onCreateView()` and update implementations
`getView()` made it unclear whether the implementation can reuse it's view or not.

We discovered that reusing the view introduces crashes when the view was added to another parent
before the previous parent has been destroyed.

Our `InjectedViewContainer` asks for the view only once when it adds it to it's child views.

Using `onCreateView()` makes it less likely that clients will return an existing view.
The name is taken from Fragments.

PIA-2912