feat(capture-sdk): Update Camera <-> Review screens navigation
The first CameraActivity instance is visible to the user only at the start. Once they go to
the review or analysis screen the first CameraActivity instance will not be shown again and it
will only return the result from the review or analysis screen to the client.

The review screen will stop if it has no pages to show.

PIA-3146