refactor(capture-sdk): Center the camera preview
The camera preview was left aligned and if the preview was wider than the screen then it
exceeded only the right side of the screen.

To be consistent with the iOS implementation we also have to center the preview. This way if it
is wider than the screen then it exceeds it on all sides.

PIA-3344