fix(capture-sdk): Fix `AnalysisInteractor` behaviour regarding analysis cancellations
Using the `FailureException` helper method we can can unwrap it from the
`CompletionException` that was caused by the failed upload and rethrow it.

Using the same helper method in the `AnalysisScreenPresenter` we can then unwrap the
`FailureException` from the `CompletionException` in which the previously thrown `FailureException`
was wrapped in.

PIA-3628