fix(capture-sdk): Show clear defaults dialog for all file types on the analysis screen
Previously only the review screen showed the clear defaults dialog for images and we showed
the review screen for images received from other apps via "open with".

Now we go directly to the analysis screen for all file types so we have to show the clear defaults
dialog for all file types and not just for PDFs.

PIA-3653