fix(capture-sdk): Fix reuse of `InjectedViewAdapter`s in multiple `InjectedViewContainer`s
Keep the association between an `InjectedViewAdapter` and an `InjectedViewContainer` in the new
`InjectedViewAdapterInstance` class. This is used to determine when an `InjectedViewContainer`
owns an adapter.

Associate the `InjectedViewAdapter` with the code it needs for configuration by using the new
`InjectedViewAdapterHolder`. This permits the adapter's view to be configured every time the view
is recreated.

PIA-3799