feat(bank-sdk): Don't require releasing capture (doing cleanup) before setting a new capture configuration
Cleanup is tied to extraction feedback sending and we want clients to only send us feedback after
the user has initiated a transaction.

By not requiring cleanup, the `GiniCapture` instance can be recreated also when users abandoned
before initiating a transaction.

Temporary images remain on disk only if the user cancelled the SDK or if there was an error.
The images will be removed when recreating the `GiniCapture` instance.
If analysis has completed, then the images will be removed before the SDK returns with the
extractions.

PIA-4178