feat(capture-sdk): Always call the cancelled result when the user leaves the host fragment via a back action
The top and bottom navigation bar back/close buttons trigger a back button press to allow the
nav graph to handle back navigation the same way for our navbars as for the system back button.

For popping the back stack a back button press is triggered as well. Only exception is the
`FileChooserFragment` where setting the fragment result didn't work when triggering the back
button press.

BSDK-258