refactor(bank-sdk): Use the `CaptureFlowFragment` in the `CaptureFlowActivity`
The `CameraActivity` is not used anymore by the `CaptureFlowActivity`.

Simplifies the `CaptureFlowContract` and the passing of input and output to and from the
`CaptureFlowActivity`. Previously the `CaptureResult` was not parcelized and returned directly from
the `CaptureFlowActivity`.

Checking that any of the Pay 5 extractions is available was moved into the
`CaptureFlowFragment` from the `CaptureFlowContract` to share the behaviour between activity and
fragment usage.

BSDK-258, BSDK-259