fix(capture-sdk): Fix camera focus indicator position
Add the ImageView for the focus indicator into a FrameLayout which takes the same size as the
CameraPreviewContainer to ensure that its position is relative to the CameraPreviewContainer.

Move the activity indicator background view in the camera screen tablet layouts out of the
CameraPreviewContainer. The CameraPreviewContainer must have only the camera preview view as its
child view.

Don't cover the top navigation bar with the activity indicator background.

PP-313