feat(bank-sdk): Replace `SplashSingleActivity` with `CaptureFlowHostActivity` in the example app
We found that `SplashSingleActivity` was testing `GiniBank.createCaptureFlowFragmentForDocument()`
which is already used under the hood by `GiniBank.startCaptureFlowForDocument()` and
`GiniBank.startCaptureFlowForIntent()`. But `GiniBank.createCaptureFlowFragmentForIntent()` was not
used by the example app. To remedy this we removed the `SplashSingleActivity` and updated the
`CaptureFlowHostActivity` to use intent-filters to launch the Bank SDK via `GiniBank.createCaptureFlowFragmentForIntent()`.

Now all six GiniBank launch methods can be tested:
1. `GiniBank.startCaptureFlow()` via the "Photopayment" button in the `MainActivity`
2. `GiniBank.startCaptureFlowForIntent()` via the `MainActivity` intent-filters labeled "Gini Bank"
3. `GiniBank.startCaptureFlowForDocument()` via the `SplashActivity` intent-filters labeled
   "Splash Activity"
4. `GiniBank.createCaptureFlowFragment()` via the "Single Activity" button in the `MainActivity`
5. `GiniBank.createCaptureFlowFragmentForIntent()` via the `CaptureFlowHostActivity` intent-filters
   labeled "Single Activity"
6. `GiniBank.createCaptureFlowFragmentForDocument()` via 2. and 3.

PP-302