refactor(health-sdk): Observe the selected payment provider app in the `ReviewFragment`
We can use the selected payment provider app's flow as the source of truth. This way we can avoid
passing the selected payment provider app to the fragment.

We also get the installed state updates directly from the PaymentComponent without having to query it.
We just ask the PaymentComponent to recheck which apps are installed and this results in the selected
payment provider app flow emitting a new instance with the updated isInstalled state.

IPC-240