refactor(health-sdk): Observe the selected payment provider app in the `InstallAppBottomSheet`
We can use the selected payment provider app's flow as the source of truth. This way we can avoid
passing the selected payment provider app to the bottom sheet.

We also get the installed state updates directly from the PaymentComponent whenever it changes
without having to explicitly query it.

IPC-240