fix(health-sdk): Use `MaterialButton` explicitly to avoid issues when clients use an AppCompat theme
When we use a `<Button>` in an xml layout, then the button is created with an inflater that depends
on the used theme. When the theme is `Theme.AppCompat.*`, then the button is created by
`AppCompatViewInflater.createButton()` and thus our SDK's Material 3 button style will contain
attributes that do not exist in the AppCompat theme.

When we use the `MaterialButton` explicitly, then we avoid relying on the theme for deciding which
button implementation is inflated.

IPC-282