refactor(core-api-library): Don't pass network security config to TrustKit in tests if running on Android N or newer
On Android N (API Level 24) or newer the network security config is used from the Android Manifest.

This allows tests to run on staging (e.g. health-api.stage.gini.net) where SSL certificates are
self-signed and tied to the specific domain (no wildcard domain for example *.stage.gini.net is used).

If the network security config is passed to TrustKit, then checking if the server is trusted on
staging fails in our `PubKeyManager.checkServerTrusted()` method.